+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix: Set current thread only if there are no exceptions
 * Enhancement: Set global tags on SentryOptions and load them from external configuration (#1066)
 * Ref: Refactor resolving SpanContext for Throwable (#1068)
+* Enhancement: Add more convenient method to start a child span (#1073)
 
 # 4.0.0-alpha.1